fix: Reafctored code and fixed issue with exatc out
