+++ src/entities/trades/velodrome/contants.ts
@@ -1,0 +1,2 @@
+export const ROUTER_ADDRESS = '0x9c12939390052919aF3155f41Bf4160Fd3666A6f'
+export const LIBRARY_ADDRESS = '0xfb1Fc21D2937bF5a49D480189e7FEd42bF8282aD'
